move tests to proper place
